`bivariate` vector

This creates a bivariate variable that allows us to represent the mean and variance of a distribution as a single variable.

Usage

bivariate(x = distributional::new_dist())

is_bivariate(x)

Arguments

x

* For `percent()`: a distribution from the `distributional` package * For `is_percent()`: An object to test.

Value

An S3 vector of class `ggdibbler_bivariate`.

Examples

bivariate(toy_temp_dist$temp_dist)
#> <ggdibbler_bivariate[99]>
#>  [1] [30,0.820] [30,1.000] [26,0.300] [23,0.690] [28,0.800] [29,2.000]
#>  [7] [28,0.130] [29,0.820] [29,0.640] [28,3.200] [26,1.500] [28,0.560]
#> [13] [26,0.540] [27,0.300] [30,0.180] [29,0.260] [27,0.470] [28,1.900]
#> [19] [28,1.500] [26,1.200] [25,0.760] [28,0.620] [29,0.540] [28,0.140]
#> [25] [27,1.900] [25,0.410] [25,0.210] [29,0.220] [26,0.200] [26,0.370]
#> [31] [27,0.310] [22,1.000] [29,0.100] [27,0.280] [27,0.690] [29,1.300]
#> [37] [25,1.400] [28,0.780] [28,1.100] [28,0.510] [26,0.640] [28,1.000]
#> [43] [28,0.640] [27,0.140] [27,0.110] [27,2.500] [27,1.100] [27,1.300]
#> [49] [26,0.460] [27,2.900] [26,1.200] [27,0.096] [30,0.760] [25,0.210]
#> [55] [26,0.620] [24,0.760] [29,0.330] [27,1.500] [26,0.380] [26,0.920]
#> [61] [29,0.850] [26,0.440] [26,0.150] [26,2.500] [27,0.240] [28,1.700]
#> [67] [27,1.200] [25,0.180] [29,1.800] [28,0.180] [27,0.880] [25,0.930]
#> [73] [28,1.300] [24,0.097] [29,0.890] [26,0.280] [27,0.560] [28,0.320]
#> [79] [29,0.410] [28,1.600] [28,1.600] [28,0.650] [28,0.560] [28,0.130]
#> [85] [28,2.400] [28,0.100] [27,0.690] [28,0.730] [25,1.000] [25,0.540]
#> [91] [27,1.900] [25,0.790] [26,1.600] [27,0.150] [25,0.210] [27,0.230]
#> [97] [27,1.800] [25,0.230] [27,0.670]