Generates a VSUP palette

This function creates a value supressed uncertainty palette so that a 1D colour palette can be transformed into a 2D palette that allows variance to be mapped to an axis. Unlike the bivariate palette, this palette merges values.

Usage

pal_vsup(colours, method = "tree", type = "desaturate", n = NULL, amount = 0.9)

Arguments

colours

character vector (of length d) of Hex colours.

method

the mixing method used to combine colour values. Can be set to either "tree" (for the binary tree mix) or "shrinkage" to have a shrinkage mix. By default it is set to "tree".

type

a character to highlight how the values will be transformed between layers of the palette. Can be one of "desaturate", "lighten", or "transparency". Is set to "desaturate" by default.

n

an integer representing the number of times the base colours should be transformed and mixed. I.e. the depth of the palette.

amount

an integer representing the amount of desaturation, lightening, or transparency between the base colour and the final colour in the palette (with 1 being completely desaturated). Can also be a numeric vector of length n that provides the amount of desaturation occuring at each layer.

Value

A character vector representing the n*d colours in the value supressed uncertainty palette.

Examples

colourvalues <- colorspace::sequential_hcl(4, palette = "YlOrRd")
pal_vsup(colourvalues)
#>      [,1]      [,2]      [,3]     
#> [1,] "#7D0025" "#984A45" "#A7A19F"
#> [2,] "#EB5500" "#984A45" "#A7A19F"
#> [3,] "#F7C252" "#F2E2BA" "#A7A19F"
#> [4,] "#FFFFC8" "#F2E2BA" "#A7A19F"